reallocate plotting-related helper functions from utils.py to utils_plot.py
